app/server/requirements.txt
Flask==0.12.2
agnostic-calculator==1.0.0
tavern==0.10.0
